+++ src/include/coin.hpp
@@ -23,6 +23,7 @@
  * You should have received a copy of the GNU General Public License
  * along with Loot Generator.  If not, see <http://www.gnu.org/licenses/>.
  */
+
 #ifndef COIN_HPP
 #define COIN_HPP
