+++ src/loot.cpp
@@ -51,51 +51,3 @@ float Loot::getValue()
     return _value;
 }
 
-
-Art::Art(std::string name, float value)
-    : Loot(name)
-{
-    setValue(value);
-}
-
-void Art::setValue(float value)
-{
-    if (value >= 0)
-    {
-        _value = round(value);
-    }
-    else
-    {
-        _value = 0;
-    }
-}
-
-std::string Art::toString()
-{
-    std::ostringstream oss;
-    oss << getName() << " art worth " << getValue() << " gp";
-    return oss.str();
-}
-
-MagicItem::MagicItem(std::string name, float value)
-    : Loot(name)
-{
-    setValue(value);
-}
-
-void MagicItem::setValue(float value)
-{
-    if (value >= 0)
-    {
-        _value = round(value);
-    }
-    else
-    {
-        _value = 0;
-    }
-}
-
-std::string MagicItem::toString()
-{
-    return getName();
-}
